rename CommonUnionType to UnionType
