+++ src/Rules/Arrays/AllowedArrayKeysTypes.php
@@ -2,20 +2,20 @@
 
 namespace PHPStan\Rules\Arrays;
 
-use PHPStan\Type\CommonUnionType;
 use PHPStan\Type\FloatType;
 use PHPStan\Type\IntegerType;
 use PHPStan\Type\NullType;
 use PHPStan\Type\StringType;
 use PHPStan\Type\TrueOrFalseBooleanType;
 use PHPStan\Type\Type;
+use PHPStan\Type\UnionType;
 
 class AllowedArrayKeysTypes
 {
 
 	public static function getType(): Type
 	{
-		return new CommonUnionType([
+		return new UnionType([
 			new IntegerType(),
 			new StringType(),
 			new FloatType(),
